:recycle: create summarizer module
